Tools update

+++ Tools/TableEntries/FIB_Entry.cs
@@ -5,9 +5,9 @@ namespace Tools.Table_Entries
     public class FIB_Entry
     {
         public IPAddress destinationIP { get; set; }
-        public ushort portOut { get; set; }
+        public int portOut { get; set; }
 
-        public FIB_Entry(IPAddress destinationIp, ushort portOut)
+        public FIB_Entry(IPAddress destinationIp, int portOut)
         {
             destinationIP = destinationIp;
             this.portOut = portOut;
